+++ changelog.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Testing
     * Locale setting unit test
     * Session settings (locale and theme) unit test
+    * Auth pages access unit tests
     
 ### Changed